Adding repool logic for CustomCapture
