+++ src/main/java/org/mskcc/limsrest/service/ToggleSampleQcStatus.java
@@ -5,6 +5,7 @@ import com.velox.api.datarecord.InvalidValue;
 import com.velox.api.datarecord.IoError;
 import com.velox.api.datarecord.NotFound;
 import com.velox.api.user.User;
+import com.velox.api.util.ServerException;
 import com.velox.sapioutils.client.standalone.VeloxConnection;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -16,6 +17,8 @@ import org.springframework.stereotype.Service;
 
 import java.rmi.RemoteException;
 import java.util.List;
+
+import static org.mskcc.util.VeloxConstants.SAMPLE;
 
 /**
  * A queued task that takes a request id and returns all some request information and some sample information
@@ -37,10 +40,11 @@ public class ToggleSampleQcStatus extends LimsTask {
     String analyst;
     String note;
     String fastqPath;
+    String recipe;
 
     private QcStatusAwareProcessAssigner qcStatusAwareProcessAssigner = new QcStatusAwareProcessAssigner();
 
-    public void init(long recordId, String status, String requestId, String correctedSampleId, String run, String qcType, String analyst, String note, String fastqPath) {
+    public void init(long recordId, String status, String requestId, String correctedSampleId, String run, String qcType, String analyst, String note, String fastqPath, String recipe) {
         this.recordId = recordId;
         this.status = status;
         this.requestId = requestId;
@@ -49,6 +53,7 @@ public class ToggleSampleQcStatus extends LimsTask {
         this.analyst = analyst;
         this.note = note;
         this.fastqPath = fastqPath;
+        this.recipe = recipe;
         if ("Post".equals(qcType)) {
             isSeqAnalysisSampleqc = false;
         }
@@ -73,10 +78,14 @@ public class ToggleSampleQcStatus extends LimsTask {
                 QcStatus qcStatus = QcStatus.fromString(status);
                 setSeqAnalysisSampleQcStatus(seqQc, qcStatus, status, user);
 
-                if (qcStatus == QcStatus.RESEQUENCE_POOL || qcStatus == QcStatus.REPOOL_SAMPLE)
+                if (qcStatus == QcStatus.RESEQUENCE_POOL){
                     qcStatusAwareProcessAssigner.assign(dataRecordManager, user, seqQc, qcStatus);
+                }
+                else if(qcStatus == QcStatus.REPOOL_SAMPLE){
+                    assignRepoolProcess(seqQc, qcStatus);
+                }
+                dataRecordManager.storeAndCommit("SeqAnalysisSampleQC updated to " + status, null, user);
 
-                dataRecordManager.storeAndCommit("SeqAnalysisSampleQC updated to " + status, user);
                 log.info("SeqAnalysisSampleQC updated to:" + status + " from:" + currentStatusLIMS);
             } else {
                 List<DataRecord> request = dataRecordManager.queryDataRecords("Request", "RequestId = '" + requestId + "'", user);
@@ -141,6 +150,107 @@ public class ToggleSampleQcStatus extends LimsTask {
         }
 
         return status;
+    }
+
+    /**
+     * Assigns process to correct sample record. Dependent upon recipe as customCapture recipes (e.g. "hemepact",
+     * "impact", & "msk-access") will need to locate the correct sample to assign the process.
+     *
+     * @param seqQc
+     * @param qcStatus
+     * @throws ServerException
+     * @throws RemoteException
+     */
+    private void assignRepoolProcess(DataRecord seqQc, QcStatus qcStatus) throws ServerException, RemoteException{
+        // If recipe is CustomCapture, assign the rePool process to the custom-capture sample record
+        String[] customRepool = new String[] {"hemepact", "impact", "msk-access"};
+        for(String type : customRepool){
+            if(this.recipe.toLowerCase().contains(type)){
+                assignCustomCaptureRepoolProcess(seqQc, qcStatus);
+                return;
+            }
+        }
+        // Otherwise, assign the process on the input sample record
+        qcStatusAwareProcessAssigner.assign(dataRecordManager, user, seqQc, qcStatus);
+    }
+
+    /**
+     * Searches for record w/ 'PoolingSampleLibProtocol' and assigns process to that record.
+     *      - Input DataRecord MUST be a customCapture Sample record.
+     *
+     * @param seqQc
+     * @param qcStatus
+     */
+    private void assignCustomCaptureRepoolProcess(DataRecord seqQc, QcStatus qcStatus) {
+        DataRecord[] childSamples = getParentsOfType(seqQc, SAMPLE);
+        if (childSamples.length > 0){
+            log.info(String.format("Found record %s. Searching for child sample with 'PoolingSampleLibProtocol'", recordId));
+            DataRecord record;
+            while (childSamples.length > 0) {
+                record = childSamples[0];
+                if (isRecordForRepooling(record)) {
+                    String pooledSampleRecord = Long.toString(record.getRecordId());
+                    log.info(String.format("Found sample, %s, with 'PoolingSampleLibProtocol'", pooledSampleRecord));
+                    qcStatusAwareProcessAssigner.assign(dataRecordManager, user, record, qcStatus);
+                    return;
+                }
+                childSamples = getChildrenOfType(record, SAMPLE);
+            }
+        }
+        log.error(String.format("Failed to assign Repool Process for %s", recordId));
+    }
+
+    /**
+     * Determines if the Sample DataRecord is the one that should have its status set. This is determiend by whether
+     * record has a child type with a 'PoolingSampleLibProtocol' set
+     *
+     * @param record
+     * @return boolean
+     */
+    private boolean isRecordForRepooling(DataRecord record) {
+        // TOOD - use com.velox.sloan.cmo.recmodels's code generator
+        DataRecord[] poolingSampleLibProtocol = getChildrenOfType(record, "PoolingSampleLibProtocol");
+        return poolingSampleLibProtocol.length > 0;
+    }
+
+    /**
+     * Returns a list of records that are children of the input record in the input table
+     *
+     * @param record
+     * @param table
+     * @return
+     */
+    private DataRecord[] getChildrenOfType(DataRecord record, String table) {
+        try {
+            return record.getChildrenOfType(table, user);
+        } catch (IoError | RemoteException e) {
+            log.error(String.format("Error getting children from %s dataType for record %s. Error: %s",
+                    table,
+                    record.getRecordId(),
+                    e.getMessage()));
+        }
+        return new DataRecord[0];
+    }
+
+    /**
+     * Returns a list of records that are parents of the input record in the input table
+     *
+     * @param record
+     * @param table
+     * @return
+     */
+    private DataRecord[] getParentsOfType(DataRecord record, String table) {
+        try {
+            List<DataRecord> dataRecords = record.getParentsOfType(table, user);
+            DataRecord[] dataRecordsArray = new DataRecord[dataRecords.size()];
+            return dataRecords.toArray(dataRecordsArray);
+        } catch (IoError | RemoteException e) {
+            log.error(String.format("Error getting parents from %s dataType for record %s. Error: %s",
+                    table,
+                    record.getRecordId(),
+                    e.getMessage()));
+        }
+        return new DataRecord[0];
     }
 
     protected static void setSeqAnalysisSampleQcStatus(DataRecord seqQc, QcStatus qcStatus, String status, User user)
